Extract ImageProcessor from ChatViewModel: saveImagesToDisk, encodeImages, padToSquare
